+++ src/Milestones/Model.php
@@ -6,8 +6,10 @@ namespace Give\Milestones;
 class Model {
 
 	// Settings
-	protected $ids;
 	protected $title;
+	protected $description;
+	protected $image;
+	protected $ids;
 
 	// Internal
 	protected $forms = [];
@@ -19,8 +21,10 @@ class Model {
 	 * @since 2.9.0
 	 **/
 	public function __construct( array $args ) {
-		isset( $args['ids'] ) ? $this->ids     = $args['ids'] : $this->ids = [];
-		isset( $args['title'] ) ? $this->title = $args['title'] : $this->title = __( 'Sample Milestone Title', 'give' );
+		isset( $args['title'] ) ? $this->title             = $args['title'] : $this->title = __( 'Sample Milestone Title', 'give' );
+		isset( $args['description'] ) ? $this->description = $args['description'] : $this->description = __( 'This is a sample description.', 'give' );
+		isset( $args['image'] ) ? $this->image             = $args['image'] : $this->image = '';
+		isset( $args['ids'] ) ? $this->ids                 = $args['ids'] : $this->ids = [];
 	}
 
 	/**
@@ -96,12 +100,30 @@ class Model {
 	}
 
 	/**
-	 * Get template path for Milestone component template
+	 * Get description for Milestone
 	 *
 	 * @return string
 	 * @since 2.9.0
 	 **/
+	protected function getDescription() {
+		return $this->description;
+	}
+
+	/**
+	 * Get image for Milestone
+	 *
+	 * @return string
+	 * @since 2.9.0
+	 **/
+	protected function getImage() {
+		return $this->image;
+	}
+
+	/**
+	 * Get template path for Milestone component template
+	 * @since 2.9.0
+	 **/
 	public function getTemplatePath() {
-		return GIVE_PLUGIN_DIR . '/src/Milestones/templates/milestone.php';
+		return GIVE_PLUGIN_DIR . '/src/Milestones/resources/views/milestone.php';
 	}
 }