+++ src/Milestones/resources/views/milestone.php
@@ -8,13 +8,18 @@
 
 <div class="give-milestone">
 	<div class="give-milestone__section">
+		<?php if ( ! empty( $this->getImage() ) ) : ?>
 		<div class="give-milestone__image">
-			<img />
+			<img src="<?php echo $this->getImage(); ?>"/>
 		</div>
+		<?php endif; ?>
 	</div>
 	<div class="give-milestone__section">
 		<div class="give-milestone__title">
 			<?php echo $this->getTitle(); ?>
+		</div>
+		<div class="give-milestone__description">
+			<?php echo $this->getDescription(); ?>
 		</div>
 		<div class="give-milestone__progress">
 			<div class="give-milestone__progress-bar" style="width: 60%"></div>